+++ data-index/data-index-storage/data-index-storage-jdbc/src/test/java/org/kie/kogito/index/jdbc/mapper/ProcessDefinitionEntityMapperIT.java
@@ -16,16 +16,12 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.index.oracle.query;
+package org.kie.kogito.index.jdbc.mapper;
 
-import org.kie.kogito.index.jpa.query.AbstractUserTaskInstanceEntityQueryIT;
+import org.kie.kogito.index.jpa.mapper.AbstractProcessDefinitionEntityMapperIT;
 
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.h2.H2DatabaseTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
 @QuarkusTest
-@QuarkusTestResource(H2DatabaseTestResource.class)
-class UserTaskInstanceEntityQueryIT extends AbstractUserTaskInstanceEntityQueryIT {
-
+class ProcessDefinitionEntityMapperIT extends AbstractProcessDefinitionEntityMapperIT {
 }
